perf(BSI): keep BatchEqual fallback in-package, drop exported helper
