+++ BitSliceIndexing/bsi_test.go
@@ -503,161 +503,49 @@ func TestTransposeWithCountsNil(t *testing.T) {
 	assert.Equal(t, int64(2), a)
 }
 
-func TestBatchEqualParallelBSIScanHelperAssertion(t *testing.T) {
-	unsortedCols := []uint32{10, 5, 20}
-	sortedCols := []uint32{5, 10, 20}
-	emptyCols := []uint32{}
-
-	t.Run("ParallelBSIScanHelper_Unsorted", func(t *testing.T) {
-		defer func() {
-			r := recover()
-			if r == nil {
-				t.Errorf("Expected ParallelBSIScanHelper to panic on unsorted cols")
+// TestBatchEqualLargeQueryValues drives the scattered-query scan path: a large
+// existence bitmap and a 128+ value query push BatchEqual past the crossover, and
+// the result is pinned to the GetValue ground truth across parallelism levels, so
+// the linear scan and its partitioning stay authoritative (subset of eBM) and exact.
+func TestBatchEqualLargeQueryValues(t *testing.T) {
+	rg := rand.New(rand.NewSource(12345))
+	for run := 0; run < 10; run++ {
+		// Large values (>= 2^20) and a big column set push past the scan crossover.
+		bsi := NewDefaultBSI()
+		numCols := rg.Intn(50000) + 120000
+		for col := 0; col < numCols; col++ {
+			if rg.Float64() < 0.8 {
+				val := rg.Int63n(100000) + 1048500
+				bsi.SetValue(uint64(col), val)
 			}
-			msg, ok := r.(string)
-			if !ok || msg != "ParallelBSIScanHelper: input cols must be sorted in ascending order" {
-				t.Errorf("Expected specific panic message, got: %v", r)
-			}
-		}()
-		_ = roaring.ParallelBSIScanHelper(unsortedCols, nil, 0, nil)
-	})
-
-	t.Run("ParallelBSIScanHelper_SortedAndEmpty", func(t *testing.T) {
-		dummyBA := []*roaring.Bitmap{roaring.NewBitmap()}
-		vals := []uint64{0, 1}
-		_ = roaring.ParallelBSIScanHelper(sortedCols, dummyBA, 1, vals)
-		_ = roaring.ParallelBSIScanHelper(emptyCols, dummyBA, 1, vals)
-	})
-}
-
-func TestBatchEqualParallelBSIScanHelperValsAssertion(t *testing.T) {
-	unsortedVals := []uint64{10, 5, 20}
-	sortedCols := []uint32{5, 10, 20}
-	dummyBA := []*roaring.Bitmap{roaring.NewBitmap()}
-
-	t.Run("ParallelBSIScanHelper_UnsortedVals", func(t *testing.T) {
-		defer func() {
-			r := recover()
-			if r == nil {
-				t.Errorf("Expected ParallelBSIScanHelper to panic on unsorted vals")
-			}
-			msg, ok := r.(string)
-			if !ok || msg != "ParallelBSIScanHelper: input vals must be sorted in ascending order" {
-				t.Errorf("Expected specific panic message, got: %v", r)
-			}
-		}()
-		_ = roaring.ParallelBSIScanHelper(sortedCols, dummyBA, 1, unsortedVals)
-	})
-}
-
-func TestBatchEqualManyBitplanes(t *testing.T) {
-	// Create a BSI with 70 bitplanes (more than 64!)
-	bsi := NewDefaultBSI()
-
-	bsi.eBM.Add(1)
-	bsi.eBM.Add(2)
-	bsi.eBM.Add(3)
-	bsi.eBM.Add(4)
-
-	bsi.bA = make([]*roaring.Bitmap, 70)
-	for i := range bsi.bA {
-		bsi.bA[i] = roaring.NewBitmap()
-	}
-
-	// Column 1: value is 1<<65 (so only plane 65 has it)
-	bsi.bA[65].Add(1)
-	// Column 2: value is 1<<3 (so only plane 3 has it)
-	bsi.bA[3].Add(2)
-	// Column 3: value is (1<<65) | (1<<3)
-	bsi.bA[65].Add(3)
-	bsi.bA[3].Add(3)
-	// Column 4: value is 1<<3
-	bsi.bA[3].Add(4)
-
-	query := []int64{8}
-
-	// Test Trie Path
-	resTrie := bsi.BatchEqual(0, query)
-	assert.True(t, resTrie.Contains(2))
-	assert.True(t, resTrie.Contains(4))
-	assert.False(t, resTrie.Contains(1))
-	assert.False(t, resTrie.Contains(3))
-
-	// Test Parallel Scan Path
-	vals := []uint64{8}
-	resScan := bsi.parallelBatchEqualScan(1, vals)
-	assert.True(t, resScan.Contains(2))
-	assert.True(t, resScan.Contains(4))
-	assert.False(t, resScan.Contains(1))
-	assert.False(t, resScan.Contains(3))
-}
-
-// TestBatchEqualExistenceAuthority pins BatchEqual results to the existence
-// bitmap. UnmarshalBinary accepts plane data that is not a subset of eBM (the
-// checked-in testdata/age fixture is such data), and every read path treats
-// eBM as authoritative, so columns present in a plane but absent from eBM must
-// never appear in results.
-func TestBatchEqualExistenceAuthority(t *testing.T) {
-	// Synthetic state: column 2 has bits in plane 0 but is absent from eBM.
-	ebm := roaring.BitmapOf(1)
-	plane := roaring.BitmapOf(1, 2)
-	ebmData, err := ebm.MarshalBinary()
-	if err != nil {
-		t.Fatal(err)
-	}
-	planeData, err := plane.MarshalBinary()
-	if err != nil {
-		t.Fatal(err)
-	}
-	bsi := NewDefaultBSI()
-	if err := bsi.UnmarshalBinary([][]byte{ebmData, planeData}); err != nil {
-		t.Fatal(err)
-	}
-	res := bsi.BatchEqual(0, []int64{1})
-	assert.True(t, res.Contains(1))
-	assert.False(t, res.Contains(2), "column 2 is not in eBM and must not match")
-
-	// The age fixture ships with plane cardinalities above the eBM cardinality;
-	// results must still be a subset of eBM.
-	large := setupLargeBSI(t)
-	if large == nil {
-		t.Skip("skipping, large BSI setup failed")
-	}
-	for _, vals := range [][]int64{{16}, {55, 57}, {0, 1, 2, 3}} {
-		res := large.BatchEqual(0, vals)
-		outside := roaring.AndNot(res, large.GetExistenceBitmap())
-		assert.True(t, outside.IsEmpty(), "BatchEqual(%v) returned %d columns outside eBM", vals, outside.GetCardinality())
-	}
-}
-
-func TestBatchEqualManyBitplanesPanicSafety(t *testing.T) {
-	// Create a BSI with 130 bitplanes (exceeding 128)
-	bsi := NewDefaultBSI()
-
-	bsi.eBM.Add(1)
-	bsi.eBM.Add(2)
-
-	bsi.bA = make([]*roaring.Bitmap, 130)
-	for i := range bsi.bA {
-		bsi.bA[i] = roaring.NewBitmap()
-	}
-
-	// Set value on plane 129
-	bsi.bA[129].Add(1)
-
-	// query with 130 scattered values (size >= 128)
-	query := make([]int64, 130)
-	for i := range query {
-		query[i] = int64(i) * 3
-	}
-
-	// This must not panic (it should safely stay on the trie walk and complete)
-	defer func() {
-		if r := recover(); r != nil {
-			t.Fatalf("BatchEqual panicked with > 128 bitplanes: %v", r)
 		}
-	}()
 
-	res := bsi.BatchEqual(0, query)
-	_ = res
+		querySize := rg.Intn(100) + 128
+		query := make([]int64, querySize)
+		for i := range query {
+			query[i] = rg.Int63n(100100) + 1048500
+		}
+
+		// Ground truth: GetValue per existing column.
+		expected := roaring.NewBitmap()
+		valMap := make(map[int64]bool)
+		for _, q := range query {
+			valMap[q] = true
+		}
+		iter := bsi.GetExistenceBitmap().Iterator()
+		for iter.HasNext() {
+			col := iter.Next()
+			val, ok := bsi.GetValue(uint64(col))
+			if ok && valMap[val] {
+				expected.Add(col)
+			}
+		}
+
+		for _, parallelism := range []int{0, 1, 2, 4} {
+			actual := bsi.BatchEqual(parallelism, query)
+			if !actual.Equals(expected) {
+				t.Fatalf("mismatch in run %d parallelism %d: expected %v, got %v", run, parallelism, expected.ToArray(), actual.ToArray())
+			}
+		}
+	}
 }
